UserEntity関連処理削除&型定義
