--- golf-auth-api/src/domains/auth/types/User.ts
@@ -1,19 +1,0 @@
-/**
- * ユーザーモデルの定義
- */
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-  password: string;
-  createdAt: Date;
-  updatedAt: Date;
-}
-
-/**
- * ユーザー作成時に必要なデータ
- */
-export interface CreateUser {
-  email: string;
-  password: string;
-}
